+++ tests/rackspacev2-salt-arch20147/settings.py
@@ -4,20 +4,20 @@ from bootmachine.settings_tests import *  # noqa
 
 
 CONFIGURATOR_MODULE = "bootmachine.contrib.configurators.salt"
-MASTER = "rackspacev2-salt-arch20139-a"
+MASTER = "rackspacev2-salt-arch20147-a"
 
 PROVIDER_MODULE = "bootmachine.contrib.providers.rackspace_openstack_v2"
 SERVERS = [
-    {"servername": "rackspacev2-salt-arch20139-a",
+    {"servername": "rackspacev2-salt-arch20147-a",
      "roles": ["salt-master"],
      "flavor": "2",
-     "image": "Arch 2013.9",
-     "distro_module": "bootmachine.contrib.distros.rackspace_arch_20139"},
-    {"servername": "rackspacev2-salt-arch20139-b",
+     "image": "Arch 2014.7 (PVHVM)",
+     "distro_module": "bootmachine.contrib.distros.rackspace_arch_20147"},
+    {"servername": "rackspacev2-salt-arch20147-b",
      "roles": ["test-additional-minions"],
      "flavor": "2",
-     "image": "Arch 2013.9",
-     "distro_module": "bootmachine.contrib.distros.rackspace_arch_20139"},
+     "image": "Arch 2014.7 (PVHVM)",
+     "distro_module": "bootmachine.contrib.distros.rackspace_arch_20147"},
 ]
 
 SSH_PUBLIC_KEY = os.path.join(os.environ["HOME"], ".ssh", "id_ecdsa.pub")